Crate gamesmanagement1_management

This documentation was generated from Games Management crate version 0.0.1. The original source code can be found on github.

Features

Handle the following Resources with ease from the central hub ...

achievements (reset, reset all, reset all for all players, reset for all players and reset multiple for all players)

applications (list hidden)

events (reset, reset all, reset all for all players, reset for all players and reset multiple for all players)

players (hide and unhide)

quests (reset, reset all, reset all for all players, reset for all players and reset multiple for all players)

rooms (reset and reset for all players)

scores (reset, reset all, reset all for all players, reset for all players and reset multiple for all players)

turn based matches (reset and reset for all players)

Everything else about the Games Management v1_management API can be found at the official documentation site.

Not what you are looking for ? Find all other google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.players().hide(...).doit()
let r = hub.players().unhide(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
gamesmanagement1_management = "0.0.1"

A complete example

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate gamesmanagement1_management;
use gamesmanagement1_management::cmn::Result;
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use gamesmanagement1_management::GamesManagement;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = GamesManagement::new(hyper::Client::new(), auth);
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.players().hide("applicationId", "playerId")
             .doit();
 
match result {
    Result::HttpError(err) => println!("HTTPERROR: {:?}", err),
    Result::MissingAPIKey => println!("Missing API Key"),
    Result::MissingToken => println!("Missing Token"),
    Result::Failure(_) => println!("General Failure (Response doesn't print)"),
    Result::FieldClash(clashed_field) => println!("FIELD CLASH: {:?}", clashed_field),
    Result::Success(_) => println!("Success (value doesn't print)"),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

About Customization/Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

About Parts

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifyable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

About Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are borrowed

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Modules

cmn

Macros

map!
This macro is advertised in the documentation, which is why we deliver it as well

Structs

AchievementMethodsBuilder
A builder providing access to all methods supported on achievement resources. It is not used directly, but through the GamesManagement hub.

AchievementResetAllForAllPlayerMethodBuilder
Resets all draft achievements for all players. This method is only available to user accounts for your developer console.

AchievementResetAllMethodBuilder
Resets all achievements for the currently authenticated player for your application. This method is only accessible to whitelisted tester accounts for your application.

AchievementResetAllResponse
This is a JSON template for achievement reset all response.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

reset all achievements (response)

AchievementResetForAllPlayerMethodBuilder
Resets the achievement with the given ID for all players. This method is only available to user accounts for your developer console. Only draft achievements can be reset.

AchievementResetMethodBuilder
Resets the achievement with the given ID for the currently authenticated player. This method is only accessible to whitelisted tester accounts for your application.

AchievementResetMultipleForAllPlayerMethodBuilder
Resets achievements with the given IDs for all players. This method is only available to user accounts for your developer console. Only draft achievements may be reset.

AchievementResetMultipleForAllRequest
This is a JSON template for multiple achievements reset all request.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

reset multiple for all players achievements (request)

AchievementResetResponse
This is a JSON template for an achievement reset response.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

reset achievements (response)

ApplicationListHiddenMethodBuilder
Get the list of players hidden from the given application. This method is only available to user accounts for your developer console.

ApplicationMethodsBuilder
A builder providing access to all methods supported on application resources. It is not used directly, but through the GamesManagement hub.

EventMethodsBuilder
A builder providing access to all methods supported on event resources. It is not used directly, but through the GamesManagement hub.

EventResetAllForAllPlayerMethodBuilder
Resets all draft events for all players. This method is only available to user accounts for your developer console. All quests that use any of these events will also be reset.

EventResetAllMethodBuilder
Resets all player progress on all events for the currently authenticated player. This method is only accessible to whitelisted tester accounts for your application. All quests for this player will also be reset.

EventResetForAllPlayerMethodBuilder
Resets the event with the given ID for all players. This method is only available to user accounts for your developer console. Only draft events can be reset. All quests that use the event will also be reset.

EventResetMethodBuilder
Resets all player progress on the event with the given ID for the currently authenticated player. This method is only accessible to whitelisted tester accounts for your application. All quests for this player that use the event will also be reset.

EventResetMultipleForAllPlayerMethodBuilder
Resets events with the given IDs for all players. This method is only available to user accounts for your developer console. Only draft events may be reset. All quests that use any of the events will also be reset.

EventsResetMultipleForAllRequest
This is a JSON template for multiple events reset all request.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

reset multiple for all players events (request)

GamesManagement
Central instance to access all GamesManagement related resource activities

GamesPlayedResource
This is a JSON template for metadata about a player playing a game with the currently authenticated user.

This type is not used in any activity, and only used as part of another schema.

GamesPlayerExperienceInfoResource
This is a JSON template for 1P/3P metadata about the player's experience.

This type is not used in any activity, and only used as part of another schema.

GamesPlayerLevelResource
This is a JSON template for 1P/3P metadata about a user's level.

This type is not used in any activity, and only used as part of another schema.

HiddenPlayer
This is a JSON template for the HiddenPlayer resource.

This type is not used in any activity, and only used as part of another schema.

HiddenPlayerList
This is a JSON template for a list of hidden players.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list hidden applications (response)

Player
This is a JSON template for a Player resource.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

hide players (none)

unhide players (none)

PlayerHideMethodBuilder
Hide the given player's leaderboard scores from the given application. This method is only available to user accounts for your developer console.

PlayerMethodsBuilder
A builder providing access to all methods supported on player resources. It is not used directly, but through the GamesManagement hub.

PlayerName
An object representation of the individual components of the player's name. For some players, these fields may not be present.

This type is not used in any activity, and only used as part of another schema.

PlayerScoreResetAllResponse
This is a JSON template for a list of leaderboard reset resources.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

reset all scores (response)

PlayerScoreResetResponse
This is a JSON template for a list of reset leaderboard entry resources.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

reset scores (response)

PlayerUnhideMethodBuilder
Unhide the given player's leaderboard scores from the given application. This method is only available to user accounts for your developer console.

QuestMethodsBuilder
A builder providing access to all methods supported on quest resources. It is not used directly, but through the GamesManagement hub.

QuestResetAllForAllPlayerMethodBuilder
Resets all draft quests for all players. This method is only available to user accounts for your developer console.

QuestResetAllMethodBuilder
Resets all player progress on all quests for the currently authenticated player. This method is only accessible to whitelisted tester accounts for your application.

QuestResetForAllPlayerMethodBuilder
Resets all player progress on the quest with the given ID for all players. This method is only available to user accounts for your developer console. Only draft quests can be reset.

QuestResetMethodBuilder
Resets all player progress on the quest with the given ID for the currently authenticated player. This method is only accessible to whitelisted tester accounts for your application.

QuestResetMultipleForAllPlayerMethodBuilder
Resets quests with the given IDs for all players. This method is only available to user accounts for your developer console. Only draft quests may be reset.

QuestsResetMultipleForAllRequest
This is a JSON template for multiple quests reset all request.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

reset multiple for all players quests (request)

RoomMethodsBuilder
A builder providing access to all methods supported on room resources. It is not used directly, but through the GamesManagement hub.

RoomResetForAllPlayerMethodBuilder
Deletes rooms where the only room participants are from whitelisted tester accounts for your application. This method is only available to user accounts for your developer console.

RoomResetMethodBuilder
Reset all rooms for the currently authenticated player for your application. This method is only accessible to whitelisted tester accounts for your application.

ScoreMethodsBuilder
A builder providing access to all methods supported on score resources. It is not used directly, but through the GamesManagement hub.

ScoreResetAllForAllPlayerMethodBuilder
Resets scores for all draft leaderboards for all players. This method is only available to user accounts for your developer console.

ScoreResetAllMethodBuilder
Resets all scores for all leaderboards for the currently authenticated players. This method is only accessible to whitelisted tester accounts for your application.

ScoreResetForAllPlayerMethodBuilder
Resets scores for the leaderboard with the given ID for all players. This method is only available to user accounts for your developer console. Only draft leaderboards can be reset.

ScoreResetMethodBuilder
Resets scores for the leaderboard with the given ID for the currently authenticated player. This method is only accessible to whitelisted tester accounts for your application.

ScoreResetMultipleForAllPlayerMethodBuilder
Resets scores for the leaderboards with the given IDs for all players. This method is only available to user accounts for your developer console. Only draft leaderboards may be reset.

ScoresResetMultipleForAllRequest
This is a JSON template for multiple scores reset all request.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

reset multiple for all players scores (request)

TurnBasedMatcheMethodsBuilder
A builder providing access to all methods supported on turnBasedMatche resources. It is not used directly, but through the GamesManagement hub.

TurnBasedMatcheResetForAllPlayerMethodBuilder
Deletes turn-based matches where the only match participants are from whitelisted tester accounts for your application. This method is only available to user accounts for your developer console.

TurnBasedMatcheResetMethodBuilder
Reset all turn-based match data for a user. This method is only accessible to whitelisted tester accounts for your application.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.