Struct gamesManagement1_management::PlayerScoreResetResponse

pub struct PlayerScoreResetResponse {
    pub definition_id: Option<String>,
    pub kind: Option<String>,
    pub reset_score_time_spans: Vec<String>,
}

This is a JSON template for a list of reset leaderboard entry resources.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

reset scores (response)

Fields

definition_id

The ID of an leaderboard for which player state has been updated.

kind

Uniquely identifies the type of this resource. Value is always the fixed string gamesManagement#playerScoreResetResponse.

reset_score_time_spans

The time spans of the updated score. Possible values are:
- "ALL_TIME" - The score is an all-time score. - "WEEKLY" - The score is a weekly score. - "DAILY" - The score is a daily score.